# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Arabic Interview Guide API",
    "version": "2.0.0",
    "description": "REST API backing the Arabic Interview Guide, a full-stack technical interview preparation app. Questions are grouped into categories and each question owns one or more answers. Features include full-text search, pagination, and difficulty/tags metadata.\n\nAll data is served from a local SQLite database. **This API has no authentication** - it is intended for local development and demonstration.",
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
...
  "paths": {
...
    "/api/stats": {
      "get": {
        "tags": ["System"],
        "summary": "Aggregate counts and per-category counts",
        "description": "Returns the total number of categories, questions and answers, plus every category with its `question_count`. Categories are ordered alphabetically by label, matching `GET /api/categories`. Drives the header summary and the per-category tab badges.",
        "responses": {
          "200": {
            "description": "Totals and per-category counts",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatsResponse" },
                "example": {
                  "total_categories": 8,
                  "total_questions": 61,
                  "total_answers": 140,
                  "categories": [
                    {
                      "id": "laravel-core",
                      "label": "Laravel Core",
                      "icon": "🔴",
                      "color": "#FF4444",
                      "question_count": 12
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "List categories",
        "description": "Returns every category ordered alphabetically by label.",
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Category" } }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Create a category",
        "description": "The `id` is a caller-supplied slug and is not auto-generated, so it can be used as a stable route segment.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryInput" },
              "example": { "id": "devops", "label": "DevOps", "icon": "⚙️", "color": "#00B8D9" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageWithId" },
                "example": { "id": "devops", "message": "Category added successfully" }
              }
            }
          },
          "400": {
            "description": "`id` or `label` missing",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Category ID and label are required" }
              }
            }
          },
          "500": {
            "description": "Database error, e.g. a duplicate `id`",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/categories/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "devops" }
      ],
      "put": {
        "tags": ["Categories"],
        "summary": "Update a category",
        "description": "The `id` is immutable; it is matched from the URL. Omitted `icon` or `color` fields fall back to `\"\"` and `\"#666666\"` respectively.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryInput" },
              "example": { "label": "DevOps & SRE", "icon": "⚙️", "color": "#00B8D9" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Message" },
                "example": { "message": "Category updated successfully" }
              }
            }
          },
          "400": {
            "description": "`label` missing",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "404": {
            "description": "No category with that id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Category not found" }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "Delete a category",
        "description": "Cascades: every question in the category and every answer on those questions is deleted too. This is enforced by `ON DELETE CASCADE` foreign keys.",
        "responses": {
          "200": {
            "description": "Category and its dependent records deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Message" },
                "example": { "message": "Category deleted successfully" }
              }
            }
          },
          "404": {
            "description": "No category with that id",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": {
            "description": "Database error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/questions": {
      "get": {
        "tags": ["Questions"],
        "summary": "List questions with their answers",
        "description": "Returns a paginated envelope of questions, each carrying its category metadata, difficulty, tags and a nested `answers` array. Answers are ordered oldest first.\n\n`page` is 1-based and `limit` defaults to 20 and is clamped to a maximum of 100; invalid values fall back to the defaults instead of erroring. Pagination counts questions, not joined rows, so a question with three answers still occupies one slot on the page.",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "description": "Restrict results to one category id.",
            "schema": { "type": "string" },
            "example": "laravel-core"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "1-based page number. Defaults to 1; values below 1 or non-numeric values fall back to 1.",
            "schema": { "type": "integer", "minimum": 1, "default": 1 },
            "example": 1
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size. Defaults to 20 and is clamped to the range 1-100.",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 },
            "example": 20
          }
        ],
        "responses": {
          "200": {
            "description": "A page of questions, each with its category metadata and answers",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedQuestions" },
                "example": {
                  "items": [
                    {
                      "id": 1,
                      "question_text": "What is the service container and how does dependency injection work?",
                      "category_id": "laravel-core",
                      "category_label": "Laravel Core",
                      "icon": "🔴",
                      "color": "#FF4444",
                      "difficulty": "intermediate",
                      "tags": ["architecture"],
                      "answers": [
                        { "id": 1, "answer_text": "The container binds class names to concrete instances." }
                      ]
                    }
                  ],
                  "total": 61,
                  "page": 1,
                  "limit": 20,
                  "pages": 4
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      },
      "post": {
        "tags": ["Questions"],
        "summary": "Create a question, optionally with answers",
        "description": "The question and its answers are inserted inside a single transaction. If any answer insert fails the whole question is rolled back, so a question is never persisted with partial answers. Accepts answers as plain strings or as objects with an `answer_text` field.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QuestionInput" },
              "example": {
                "category_id": "laravel-core",
                "question_text": "Explain the service container.",
                "answers": ["It resolves and injects dependencies.", "Bindings are registered in a service provider."]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Question created, with the new id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageWithId" },
                "example": { "id": 62, "message": "Question and answers added successfully" }
              }
            }
          },
          "400": {
            "description": "`category_id` or `question_text` missing",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Category ID and question text are required" }
              }
            }
          },
          "500": {
            "description": "Database error, e.g. an unknown `category_id`",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Full-text search across questions and answers",
        "description": "Returns the same paginated envelope as `GET /api/questions`, ordered so that questions whose own text matched rank above questions that matched only through an answer.\n\nSearch runs against SQLite FTS5 over both question text and answer text. Input is sanitised before matching - punctuation is stripped, tokens are quoted so FTS operators are inert, at most 8 tokens are used, and the final token is prefix-matched so results narrow as you type. Multi-word queries are ANDed and fall back to OR when AND yields nothing, which is needed for Arabic definite articles.\n\nA missing, empty or whitespace-only `q` is not an error: the response is 200 with an empty `items` array and `total` 0. Results are not filtered by category.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Search term. An empty or whitespace-only value returns an empty result set with a 200 response rather than an error.",
            "schema": { "type": "string" },
            "example": "dependency injection"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "1-based page number. Defaults to 1; values below 1 or non-numeric values fall back to 1.",
            "schema": { "type": "integer", "minimum": 1, "default": 1 },
            "example": 1
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size. Defaults to 20 and is clamped to the range 1-100.",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 },
            "example": 20
          }
        ],
        "responses": {
          "200": {
            "description": "A page of matching questions, ranked by relevance",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedQuestions" },
                "example": {
                  "items": [
                    {
                      "id": 1,
                      "question_text": "What is the service container and how does dependency injection work?",
                      "category_id": "laravel-core",
                      "category_label": "Laravel Core",
                      "icon": "🔴",
                      "color": "#FF4444",
                      "difficulty": "intermediate",
                      "tags": ["architecture"],
                      "answers": [
                        { "id": 1, "answer_text": "The container binds class names to concrete instances." }
                      ]
                    }
                  ],
                  "total": 3,
                  "page": 1,
                  "limit": 20,
                  "pages": 1
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/questions/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }
      ],
      "put": {
        "tags": ["Questions"],
        "summary": "Update a question and its answers",
        "description": "When `answers` is present the existing answers are deleted and the supplied list is inserted instead, inside one transaction. The `answers` array is therefore the full desired end state, not a patch. Omitting `answers` updates the question text and category only and leaves existing answers untouched.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QuestionUpdateInput" },
              "example": {
                "question_text": "Explain the service container.",
                "category_id": "laravel-advanced",
                "answers": ["Updated answer text."]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Message" },
                "example": { "message": "Question and answers updated successfully" }
              }
            }
          },
          "400": {
            "description": "`category_id` or `question_text` missing",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Category ID and question text are required" }
              }
            }
          },
          "404": {
            "description": "No question with that id",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": {
            "description": "Database error, e.g. an unknown `category_id`",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      },
      "delete": {
        "tags": ["Questions"],
        "summary": "Delete a question",
        "description": "Cascades to the question's answers.",
        "responses": {
          "200": {
            "description": "Question and its answers deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Message" },
                "example": { "message": "Question deleted successfully" }
              }
            }
          },
          "404": {
            "description": "No question with that id",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": {
            "description": "Database error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/answers/{questionId}": {
      "get": {
        "tags": ["Answers"],
        "summary": "List answers for a question",
        "description": "Returns raw answer rows ordered by id. The question list endpoint already nests answers, so this is mainly a direct-access convenience route.",
        "parameters": [
          { "name": "questionId", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }
        ],
        "responses": {
          "200": {
            "description": "Answers belonging to the question",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Answer" } }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "Caller-supplied slug used as the primary key.", "example": "laravel-core" },
          "label": { "type": "string", "description": "Display name shown on the tab.", "example": "Laravel Core" },
          "icon": { "type": "string", "description": "Emoji rendered beside the label.", "example": "🔴" },
          "color": { "type": "string", "description": "Hex colour used for the active tab, counters and answer borders.", "example": "#FF4444" },
          "question_count": { "type": "integer", "description": "Number of questions in this category. Counted by a LEFT JOIN, so categories with no questions report 0.", "example": 12 }
        }
      },
      "CategoryInput": {
        "type": "object",
        "required": ["label"],
        "properties": {
          "id": { "type": "string", "description": "Required on create, ignored on update.", "example": "devops" },
          "label": { "type": "string", "example": "DevOps" },
          "icon": { "type": "string", "example": "⚙️" },
          "color": { "type": "string", "example": "#00B8D9" }
        }
      },
...
      "StatsResponse": {
        "type": "object",
        "properties": {
          "total_categories": { "type": "integer", "example": 8 },
          "total_questions": { "type": "integer", "example": 61 },
          "total_answers": { "type": "integer", "example": 140 },
          "categories": {
            "type": "array",
            "description": "Every category with its `question_count`, ordered by label.",
            "items": { "$ref": "#/components/schemas/Category" }
          }
        }
      },
...
      "Message": {
        "type": "object",
        "properties": { "message": { "type": "string", "example": "Question deleted successfully" } }
      },
      "MessageWithId": {
        "type": "object",
        "properties": {
          "id": { "oneOf": [{ "type": "integer" }, { "type": "string" }], "example": 62 },
          "message": { "type": "string", "example": "Question added successfully" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string", "example": "Category not found" } }
      }
    }
  }
}